+++ os_kernel/ipc-os.h
@@ -1,6 +1,6 @@
 /* SPDX-License-Identifier: BSD-3-Clause */
 /*
- * Copyright 2018,2021 NXP
+ * Copyright 2018,2021,2024 NXP
  */
 #ifndef IPC_OS_H
 #define IPC_OS_H
@@ -24,13 +24,13 @@
 struct ipc_shm_cfg;
 
 /* function declarations */
-int ipc_os_init(const uint8_t instance, const struct ipc_shm_cfg *cfg,
-		int (*rx_cb)(const uint8_t, int));
+int8_t ipc_os_init(const uint8_t instance, const struct ipc_shm_cfg *cfg,
+		uint32_t (*rx_cb)(const uint8_t, int));
 void ipc_os_free(const uint8_t instance);
 uintptr_t ipc_os_get_local_shm(const uint8_t instance);
 uintptr_t ipc_os_get_remote_shm(const uint8_t instance);
 void *ipc_os_map_intc(void);
 void ipc_os_unmap_intc(void *addr);
-int ipc_os_poll_channels(const uint8_t instance);
+int8_t ipc_os_poll_channels(const uint8_t instance);
 
 #endif /* IPC_OS_H */
